backend/requirements.txt
# FastAPI and Server
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.9

# Database
sqlalchemy>=2.0.45
# psycopg2-binary==2.9.9  # PostgreSQL adapter (optional - only needed for PostgreSQL)
alembic==1.12.1  # Database migrations

# Environment & Config
python-dotenv==1.0.0
pydantic>=2.9.0
pydantic-settings>=2.1.0

# AI Models
groq==0.33.0
openai==2.3.0
requests==2.31.0

# Trading APIs
kiteconnect==4.2.0

# Date/Time utilities
python-dateutil==2.8.2

# CORS is handled by FastAPI's CORSMiddleware (no separate package needed)

# Utilities
httpx==0.25.2
pandas>=2.1.4  # For Excel file parsing
openpyxl>=3.1.2  # Excel file support for pandas
apscheduler==3.10.4  # For scheduled tasks (daily snapshots)


